Adjust time data for ingest attempts
A time and duration is stored in the database for each ingest attempts.
However, initial representation of this data turned out to be rather
poorly chosen, attempt time was represented as Unix time, duration as
seconds.  I changed attempt time to ISO date format and expressed
duration of an attempt in milliseconds.