Do not allow for setting incorrect file status
Ingester allow a user to select a files which it will process based on
their status.  However, this status cannot be set to value which
represent status of files which have been ingested successfully
(currently: SUCCESS) as there will be no way for the Ingester to tell
the difference between "old" and "new" successful attempts.  Thus I
added a safeguard which will raise an exception when it happens.