Pre-process exception in ingest plugins
Any exceptions occurring during an execution of a LSST task where passed
directly from the plugin to the Ingester.  Thus adding any
"pre-processing" of these error (for example, for logging purposes)
would implicitly introduced an undesired coupling between the Ingester
and a plugin implementations.  To prevent this from happening, any error
handling is now done on the plugin.  Then a plugin raises a RuntimeError
which is intercepted by the Ingester and provided error is simply logged
and/or inserted to the database.