Improve generator parsing rsync logs
Existing implementation of the generator responsible for parsing rsync
logs in search for files is transferred was to simplistic.  Firstly, it
was parsing entire log directory. Secondly, it had no mechanism allowing
to distinguish between logs that were already parsed and new ones.  I
addressed the first shortcoming by adding options which allows one to
select which subdirectory in the main log file directory needs to be
monitored.  Also, after processing a log file it now creates an empty
"guard file" indicating that a given log file was already processed
during previous iteration.