Fix issue with creating entries for new files
Turned out SQLAlchemy query result is a tuple even if the query returns
values form a single column.  As a result, while creating database
entries for new files the field 'files_id' was being assigned the
1-element tuple instead of the expected scalar value.  This was causing
problems while using SQLlite as the database back-end, but strangely
enough not in PostgreSQL.  Fixed the issue.