Remove relationship between File and Event ORMs
Following SQLAlchemy docs, I defined an explicit relationship between
File and Event ORMs.  As a result the File and in turn the Finder itself
became tightly coupled with whatever table the Event represented when
the Finder was started.  It would render this relationship meaningless
for any Ingester using different file events table than the one the
Finder was associated with so I removed it.