Validate plugin's configuration
Different ingest plugins can (and most likely will) have different
configurations.  Because of that they were not validated as I didn't
know how to capture this differences in the JSON schema I was using to
validate configuration files.  It turned out JSON schema support
applying subschemas conditionally, so it's no longer the case.

Allow the date for search method be null