Allow action specification to be null
When defining actions for Finder an operator needs to always specify
them, even if no action is really required.  In such a case program
expected providing a name of the no-op action, a null object called
Null.  That might cause some confusion with the null value (null).  I
modified a program to accept the regular null value instead of the name
of the actual null object.