Deal gracefully with database disconnections
The SQLAlchemy connects only to the database only if it is asked to
perform a task against the database.  Taking this fact into account, it
turned out that some of the try-except blocks were put in completely wrong
places.  As a result the code weren't handled gracefully disconnections
with the database, e.g., the Ingester was ceasing to operate if the
connection was lost.  I relocated the try-except blocks to make them
guard the code which actually connects to the database to prevent this
behavior in the future.