Handle empty files better
Empty files were not handled optimally by the endpoint manager.
Firstly, all empty files, except the one which was discovered first,
were treated as duplicates and deleted.  Secondly, the manager still
tried to ingest the empty file to the database system. It was causing
rather obscure errors coming from the ingest software being recorded as
the reason behind the failure.  To handle such files better, I slightly
altered the logic which tells Finder if a file is a duplicate or not and
prevented Ingester from making ingest attempts for empty files.  Now it
creates an entry of a failed attempt with a clear message about file
being empty.