Ignore incomplete rsync logs
Writing rsync transfer logs in not an atomic operation, but Finder
search plugin wasn't able to distinguish between partially and fully
transferred files.  To address this situation I modified the plugin to
make it ignore logs which were modified within a given time span and
added a guard which checks if a log file was not modified after being
parsed.