Rename no-op action
To prevent Finder for doing anything with the files it discovers one
needs to specify a no-op action in each Finder's action slot in its YAML
configuration file.  This action was named Null which could easily be
confused with YAML "null" value.  Hence I changed the name of the action
to Noop.