Make endpoint manager monitor future day(s)
Files from a single observation night can be placed in different
directories and these directories can be created in a time zone
different from the one the endpoint site operates in.  Though the
endpoint manager was already able to monitor past days, it is not able
to look "into the future" which may be required for some time zones.  To
address this issue I added an option to the search function parsing
rsync logs which will allow one to instruct the manager to "look ahead"
of local time for a given number of days (1, by default).