Add a backfill command
DBB endpoint buffer manager didn't have the capability to backfill
database tables it uses with entries regarding files which already
existed in the storage area before it was deployed.  To address this
limitation I added a new command, backfill, which allow an operator to
populate the file and event tables with entries related to these files.

At the moment, the command does *not* verify if the files are really
ingested to a Butler dataset repository.  It merely makes the manager
"aware of" of these files by making an entries in file and event tables.