Get rid of id query
The backfill tool was making an explicit database query to find out the
id of the file record it added in the earlier step.  I replace it
SQLAlechmy's session.flush() which pushes session changes to the
database's transaction buffer and allow to get id of the added file
record without making a separate query.