Simplify communication between threads
The provide some degree of scalability the Ingester is able to use a
pool of ingest threads.  However, the communication between the main
thread and the ingest threads was somewhat lacking:

* The input queue required messages to be just paths to files.  As
  a result, passing any other useful pieces of information about a
  file was not possible.  As a result, maintaining auxiliary data
  structures was necessary to compile a valid database record from an
  input message and the corresponding output message.

* There were two output queues which were only used to determine
  whether an ingest attempt was a success or a failure.

To address these shortcomings, I introduced two new data structures
representing the input and the output messages:

1. a request for an ingest attempt which holds any piece of
   information about the file for which the ingest attempt is
   requested,

2. a reply which captures outcome of a ingest attempt and any other data
   required to create a corresponding database entry.

Such an approach no longer requires maintaining the auxiliary data
structures and reduced number of message queues to two.