Allow ingester to be more selective about files
Ingester makes ingest attempts for all files that are tracked and
had no mechanism to exclude certain files which should not be ingested
to a Butler repository.  To address this shortcoming and make tracking
files easier I implemented a filtering mechanism which will allow an
operator to ignore select files based on their filenames.