Save outcome of the first fail check
For files failing multiple checks only the outcome of the last check was
persisted in the database.  It led to potentially confusing situation
when events where created for files which were being filtered out by
either the include or the exclude list.  To change this behavior I made
the ingester skip any checks after the first which a file fails to pass.
Taking into account that matching the filename against the patterns on
both lists is made first it should ensure that there will be no events,
other than about being ignored for files that are being filtered out.